Added migration to/from pytorch LSTMCell/LSTM
